Improved integration test file handling

+++ tests/autotuning_methodology/integration/test_run_experiment.py
@@ -15,11 +15,17 @@ from autotuning_methodology.experiments import (
 mockfiles_path_root = Path("tests/autotuning_methodology/integration/mockfiles/")
 mockfiles_path_source = mockfiles_path_root / "mock_gpu.json"
 mockfiles_path = Path("..") / mockfiles_path_root
-cached_visualization_path = Path(
-    "cached_data_used/visualizations/test_run_experiment/mocktest_kernel_convolution/mock_GPU_random_sample_10_iter.npz"
-)
+cached_visualization_path = Path("cached_data_used/visualizations/test_run_experiment/mocktest_kernel_convolution")
+cached_visualization_file = cached_visualization_path / "mock_GPU_random_sample_10_iter.npz"
 normal_cachefiles_path = Path("cached_data_used/cachefiles/mocktest_kernel_convolution")
 normal_cachefile_destination = normal_cachefiles_path / "mock_gpu.json"
+
+
+def _remove_dir(path: Path):
+    assert path.exists()
+    for sub in path.iterdir():
+        sub.unlink()
+    path.rmdir()
 
 
 def setup_module():
@@ -29,15 +35,18 @@ def setup_module():
     assert normal_cachefiles_path.exists()
     normal_cachefile_destination.write_text(mockfiles_path_source.read_text())
     assert normal_cachefile_destination.exists()
+    cached_visualization_path.mkdir(parents=True, exist_ok=True)
+    assert cached_visualization_path.exists()
 
 
 def teardown_module():
     # remove files where necessary
-    normal_cachefile_destination.unlink()
-    assert normal_cachefiles_path.exists()
-    for sub in normal_cachefiles_path.iterdir():
-        sub.unlink()
-    normal_cachefiles_path.rmdir()
+    if normal_cachefile_destination.exists():
+        normal_cachefile_destination.unlink()
+    _remove_dir(normal_cachefiles_path)
+    if cached_visualization_file.exists():
+        cached_visualization_file.unlink()
+    _remove_dir(cached_visualization_path)
 
 
 def test_CLI_input():
@@ -77,21 +86,25 @@ def test_run_experiment_bad_kernel_path():
         execute_experiment(experiment_filepath, profiling=False)
 
 
+@pytest.fixture(scope="session")
 def test_run_experiment():
     """Run a dummy experiment"""
     assert normal_cachefile_destination.exists()
-    if cached_visualization_path.exists():
-        cached_visualization_path.unlink()
-    assert not cached_visualization_path.exists()
+    if cached_visualization_file.exists():
+        cached_visualization_file.unlink()
+    assert not cached_visualization_file.exists()
     experiment_filepath = str(mockfiles_path / "test.json")
     (experiment, strategies, results_descriptions) = execute_experiment(experiment_filepath, profiling=False)
     validate_experiment_results(experiment, strategies, results_descriptions)
 
 
+@pytest.mark.usefixtures("test_run_experiment")
 def test_cached_experiment():
     """Retrieve a cached experiment"""
+    assert normal_cachefiles_path.exists()
     assert normal_cachefile_destination.exists()
     assert cached_visualization_path.exists()
+    assert cached_visualization_file.exists()
     experiment_filepath = str(mockfiles_path / "test.json")
     (experiment, strategies, results_descriptions) = execute_experiment(experiment_filepath, profiling=False)
     validate_experiment_results(experiment, strategies, results_descriptions)
